Improved stats history trimming, removed debounce
The debounce trimming function was causing it to spawn many functions running over top of each other causing CPU to spike. This change is still fairly CPU intensive, but we need to change the strategy for how stats are stored in LevelDb. Scanning all of the keys for every trim is slow.